pyproject.toml
[project]
name = "psychofauna"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.4",
    "onnxruntime>=1.20.1",
    "pandas>=2.2.3",
    "torch>=2.5.1",
    "transformers>=4.48.1",
]
